# Conflicts:
#	integrationTests/multiShard/block/executingMiniblocksSc_test.go
#	integrationTests/testProcessorNode.go

+++ consensus/spos/consensusState.go
@@ -20,7 +20,7 @@ type ConsensusState struct {
 	BlockBody data.BodyHandler
 	Header    data.HeaderHandler
 
-	RoundIndex     int32
+	RoundIndex     int64
 	RoundTimeStamp time.Time
 	RoundCanceled  bool
 
@@ -189,7 +189,7 @@ func (cns *ConsensusState) CanDoSubroundJob(currentSubroundId int) bool {
 }
 
 // CanProcessReceivedMessage method returns true if the message received can be processed and false otherwise
-func (cns *ConsensusState) CanProcessReceivedMessage(cnsDta *consensus.Message, currentRoundIndex int32,
+func (cns *ConsensusState) CanProcessReceivedMessage(cnsDta *consensus.Message, currentRoundIndex int64,
 	currentSubroundId int) bool {
 	if cns.IsNodeSelf(string(cnsDta.PubKey)) {
 		return false
