+++ integrationTests/multiShard/block/executingMiniblocksSc_test.go
@@ -66,7 +66,7 @@ func TestProcessWithScTxsTopUpAndWithdrawOnlyProposers(t *testing.T) {
 	fmt.Println("Delaying for nodes p2p bootstrap...")
 	time.Sleep(stepDelay)
 
-	round := uint32(0)
+	round := uint64(0)
 	round = incrementAndPrintRound(round)
 
 	initialVal := big.NewInt(10000000)
@@ -177,7 +177,7 @@ func TestProcessWithScTxsJoinAndRewardTwoNodesInShard(t *testing.T) {
 	fmt.Println("Delaying for nodes p2p bootstrap...")
 	time.Sleep(stepDelay)
 
-	round := uint32(0)
+	round := uint64(0)
 	round = incrementAndPrintRound(round)
 
 	initialVal := big.NewInt(10000000)
@@ -339,7 +339,7 @@ func TestShouldProcessWithScTxsJoinNoCommitShouldProcessedByValidators(t *testin
 	)
 }
 
-func incrementAndPrintRound(round uint32) uint32 {
+func incrementAndPrintRound(round uint64) uint64 {
 	round++
 	fmt.Printf("#################################### ROUND %d BEGINS ####################################\n\n", round)
 
@@ -364,7 +364,7 @@ func deployScTx(
 
 func proposeBlockWithScTxs(
 	nodes []*integrationTests.TestProcessorNode,
-	round uint32,
+	round uint64,
 	idxProposers []int,
 ) {
 
@@ -388,7 +388,7 @@ func syncBlock(
 	t *testing.T,
 	nodes []*integrationTests.TestProcessorNode,
 	idxProposers []int,
-	round uint32,
+	round uint64,
 ) {
 
 	fmt.Println("All other shard nodes sync the proposed block...")
